+++ flutter_searchbox_ui/example/lib/main.dart
@@ -107,7 +107,7 @@ class HomePage extends StatelessWidget {
             data: Theme.of(context).copyWith(
               // Set the transparency here
               canvasColor: Colors.white.withOpacity(
-                  .6), //or any other color you want. e.g Colors.blue.withOpacity(0.5)
+                  .8), //or any other color you want. e.g Colors.blue.withOpacity(0.5)
             ),
             child: Container(
               color: Colors.transparent,
@@ -119,27 +119,29 @@ class HomePage extends StatelessWidget {
                   child: RangeInput(
                     key: const Key("key2"),
                     id: 'range',
-                    title: "Range",
-                    rangeLabel: "until",
+                    title: "Range L",
+                    rangeLabel: "to",
                     dataField: 'original_publication_year',
                     range: const RangeType(
-                        start: 3000, end: ['other', 1990, 2000, 2010]),
+                        start: 3000, end: ['other', 1990, 2000, 2010],),
+                    defaultValue: const DefaultValue(start: 1980, end: 2060),
                     rangeLabels: RangeLabelsType(
                       start: (value) {
-                        return '# $value €';
+                        return value == 'other' ? 'Custom Other' : 'yr $value';
                       },
                       end: (value) {
-                        return '# $value €';
+                        return value == 'other' ? 'Custom Other' : 'yr $value';
                       },
                     ),
                     validateRange: (start, end) {
+                      print('custom validate');
                       if (start < end) {
                         return true;
                       }
                       return false;
                     },
                     errorMessage: (start, end) {
-                      return 'Starting value: $start seems greater than Ending Value: $end';
+                      return 'Custom error $start > $end';
                     },
                   ),
                 ),