+++ flutter_searchbox_ui/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.1-alpha] - 30-09-2021.
+
+- Fix `RangeInput` widget for bugs.
+    - validation bugs.
+    - display range labels for text input also.
+    - updated documentation with relevant example snippets.
 ## [1.0.0-alpha] - 30-09-2021.
 
 - Add `RangeInput` widget